Attempt to implement peripherals, rust gets in the way
Rust can't prove that process_tick is safe (since the tick method could
theoretically modify the Vec/Map we're iterating over), so this doesn't
build. (We don't do that, so the code *is* safe, there's just no way to
tell Rust's borrow checker that)

All the workarounds make this code more complex and difficult to
understand. The cleanest from my current perspective would be to simply
launder the reference through an unsafe pointer, but that would be
potentially very confusing from a rust perspective (and is technically
un-sound).

I suspect this may be the end of my Rust experiment.